requirements.txt
alembic==1.12.0
bcrypt==4.0.1
blinker==1.6.2
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.3.0
click==8.1.7
cryptography==41.0.3
Flask==2.3.3
Flask-Bcrypt==1.0.1
Flask-CKEditor==0.4.6
Flask-Login==0.6.2
Flask-Migrate==4.0.5
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.1.1
greenlet==2.0.2
idna==3.4
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.3
mysql-connector==2.2.9
mysql-connector-python==8.1.0
mysql-connector-python-rf==2.2.2
packaging==23.2
pluggy==1.3.0
protobuf==4.21.12
pycparser==2.21
PyMySQL==1.1.0
pytest==7.4.2
python-dotenv==1.0.0
PyYAML==6.0.1
requests==2.31.0
responses==0.23.3
SQLAlchemy==2.0.20
types-PyYAML==6.0.12.12
typing_extensions==4.7.1
urllib3==2.0.6
Werkzeug==2.3.7
WTForms==3.0.1
